add security checks for changing the appearance of some panels

+++ admin/class-accordion-slider-admin.php
@@ -199,6 +199,7 @@ class BQW_Accordion_Slider_Admin {
 				'id' => $id,
 				'lad_nonce' => wp_create_nonce( 'load-accordion-data' . $id ),
 				'sa_nonce' => wp_create_nonce( 'save-accordion' . $id ),
+				'cp_nonce' => wp_create_nonce( 'close-panel' ),
 				'remove_custom_css_js_warning' => __( 'Are you sure you want to remove the existing custom CSS and/or JavaScript? <br/> Only do this after you\'ve copied the existing code in another place.', 'accordion-slider' ),
 				'no_image' => __( 'Click to add image', 'accordion-slider' ),
 				'posts_panels' => __( 'Posts panels', 'accordion-slider' ),
@@ -1239,6 +1240,12 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.0.0
 	 */
 	public function ajax_getting_started_close() {
+		$nonce = $_POST['nonce'];
+
+		if ( ! wp_verify_nonce( $nonce, 'close-panel' ) || ! current_user_can( 'manage_options' ) ) {
+			die( 'This action was stopped for security purposes.' );
+		}
+
 		update_option( 'accordion_slider_hide_getting_started_info', true );
 
 		die();
@@ -1250,6 +1257,12 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.9.0
 	 */
 	public function ajax_close_custom_css_js_warning() {
+		$nonce = $_POST['nonce'];
+
+		if ( ! wp_verify_nonce( $nonce, 'close-panel' ) || ! current_user_can( 'manage_options' ) ) {
+			die( 'This action was stopped for security purposes.' );
+		}
+
 		update_option( 'accordion_slider_hide_custom_css_js_warning', true );
 
 		delete_option( 'accordion_slider_custom_css' );
@@ -1266,6 +1279,12 @@ class BQW_Accordion_Slider_Admin {
 	 * @since 1.8.0
 	 */
 	public function ajax_close_image_size_warning() {
+		$nonce = $_POST['nonce'];
+
+		if ( ! wp_verify_nonce( $nonce, 'close-panel' ) || ! current_user_can( 'manage_options' ) ) {
+			die( 'This action was stopped for security purposes.' );
+		}
+
 		update_option( 'accordion_slider_hide_image_size_warning', true );
  
 		die();
